[webui] Move Kiwi Repository functionality
Move Kiwi Repository edit/update/index functionality to Image controller. This
will allow us to update all the elements of an Image (repositories, packages,
etc.) at the same time.

Pair-programmed by @DavidKang and @Ana06.
