+++ src/api/app/views/webui/kiwi/_tabs.html.erb
@@ -8,7 +8,7 @@
 
 <div class="box-header header-tabs" id="package_tabs">
   <ul>
-    <%= tab 'kiwi', 'Kiwi', controller: 'kiwi/repositories', action: 'index', selected: true %>
+    <%= tab 'kiwi', 'Kiwi', controller: 'kiwi/images', action: 'show', selected: true %>
     <% if @package.name == "patchinfo" %>
       <%= tab 'details', 'Details', controller: '/webui/patchinfo', action: 'show' %>
     <% end %>
