+++ src/api/app/views/webui/kiwi/images/_repository_entry.html.haml
