+++ src/api/app/views/webui/kiwi/images/_repository_fields.html.haml
