+++ src/api/spec/cassettes/Webui_Kiwi_ImagesController/GET_import_from_package/without_a_kiwi_file/1_1_1_1.yml
@@ -186,4 +186,43 @@ http_interactions:
         </package>
     http_version: 
   recorded_at: Thu, 29 Jun 2017 14:46:53 GMT
+- request:
+    method: put
+    uri: http://localhost:3200/source/fake_project/package_without_kiwi_file/_meta?user=tom
+    body:
+      encoding: UTF-8
+      string: |
+        <package name="package_without_kiwi_file" project="fake_project">
+          <title>If Not Now, When?</title>
+          <description>Eum aliquid numquam enim.</description>
+        </package>
+    headers:
+      Accept-Encoding:
+      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
+      Accept:
+      - "*/*"
+      User-Agent:
+      - Ruby
+  response:
+    status:
+      code: 200
+      message: OK
+    headers:
+      Content-Type:
+      - text/xml
+      Cache-Control:
+      - no-cache
+      Connection:
+      - close
+      Content-Length:
+      - '167'
+    body:
+      encoding: UTF-8
+      string: |
+        <package name="package_without_kiwi_file" project="fake_project">
+          <title>If Not Now, When?</title>
+          <description>Eum aliquid numquam enim.</description>
+        </package>
+    http_version: 
+  recorded_at: Tue, 11 Jul 2017 13:53:23 GMT
 recorded_with: VCR 3.0.3
